Stop using `plot-device%` in untyped code
The class was used in snips to draw message and selection boxes. The
class contract *memory* overhead was about 6MB per frame, to create
a `plot-device%` instance and make less than 20 method calls. The
majority of costs were probably for checking the contracts for the
`DC<%>` instance passed as an init value to `plot-device%`, which
those methods used for drawing.