Use Ryan Culpepper's hack to avoid (Instance DC<%>) contract in plot/dc
Using the TR-generated contracts for plot/dc and plot3d/dc makes
them too slow for one of their main purposes: animation. The problem
is the contract generated for the (Instance DC<%>) argument. If the
argument is passed in from untyped code, the contract wrapper performs
expensive checks at every dc<%> method invocation.

This lovely hack captures the identifiers during expansion time and
exports them in a phase-1 box, bypassing the contract boundary. Using
the identifiers in untyped code would be unsafe, so they're reexported
with a contract that corresponds with the type.

When TR gains `provide/unsafe`, it should replace this hack directly.

When class contracts are fast enough to handle dc<%> instances, this
hack or its `provide/unsafe` replacement should be removed.