Provide plot3d interface unsafely to avoid excessive overhead.
These three provided functions don't take any higher-order
arguments and their results are actually implemented in untyped
code, so there should not be a risk of segfaults, plus they
check their first-order arguments. Additionally, the result
contracts have not been checked in past releases due to a bug
in contract generation, increasing confidence that this is safe.

The `get-renderer-list` function is given a wider domain so that
its use in plot3d functions is not unsafe when the contracts are
omitted.

Fixes racket/gui#86. Related to racket/typed-racket#599.