Extend `rectangles` renderer to accept open intervals
When #f is specified to one of the interval bounds of the `rectangles`
renderer, the rectangle will extend to the edge of the plot area in the
direction of that bound.  Example:

```racket
#lang racket
(require plot)
(plot (rectangles (list (vector (ivl 2 3) (ivl #f #f))
                        (vector (ivl #f #f) (ivl 2 3))
                        ;; Vertical extent from 0 to the top
                        (vector (ivl 5 6) (ivl 0 #f)))
                  #:line-style 'transparent
                  #:alpha 0.5)
      #:x-min -10 #:x-max 10
      #:y-min -10 #:y-max 10)
```

Implementation note: the rectangles renderer code used to compute the bounds
of the plot based on the bounds of the rectangles, this code is unnecessary
and complicates the case where the intervals are open-ended.  It was removed.