2d-plot-snip% -- only pass the area when creating the object
The 2d-plot-snip% object needs to obtain the plot area and convert from dc to
plot coordinates and vice-versa.  Previously, both the plot area and a convert
function were passed in to the object, but it is simpler to just pass the area
object itself and let the snip use it, as this reduces the number of
parameters for several intermediate functions.