Capture plot parameters in plot-frame and plot3d-frame (#45)
plot-frame and plot3d-frame delays the creation of the plot snip object until the size of the canvas is known. This means that the snip creation code is executed outside any parameterize scope and the values of any plot parameters (like plot-decorations?) would be ignored unless they are saved using plot-parameters and restored using parameterize/group

This fix is based on a problem reported here: https://groups.google.com/forum/#!topic/racket-users/ta-245wAYco

The issue was introduced when #39 was fixed, since at that time the plot snip was no longer created withing the scope of the parameterize call.