Allow plot-{x,y,z}-far-tick-labels? to have three options (issue #46)
The plot-{x,y,z}-far-tick-labels? parameters can have three possible values:
'auto, #t or #f. Only #t and #f are documented for users.

When a parameter is set to 'auto, the behavior is that far labels are present
only if the ticks on the far axis are different from the ticks on the near
axis (this was the plot behavior before this parameter was introduced).

When a parameter is set to #t or #f, the far labels are shown or not
respectively.  Note that labels will only be displayed if the axis ticks are
drawn on the far axis.