Add color map support for selecting colors for renderers
Plot renderers have a `#:color` argument which specifies the color to use for
drawing, this can be a `color%` object, a RGB triplet, or an integer
representing an index into a color palette.  This commit introduces a
mechanism to select which color palette (called a color map) to use when
integers are used for the `#:color` argument of the renderers.

```racket
(require plot)
(parameterize ([plot-pen-color-map 'set1]
               [line-width 3])
    (plot (list (function sin -5 5 #:color 0)
                (function cos -5 5 #:color 1))))
```

The following new API functions are introduced and updated:

* `plot-pen-color-map` and `plot-brush-color-map` are new parameters which
  specify the name of a color map to use for the `#:color` argument of
  renderer functions.

* the `->pen-color` and `->brush-color` functions have been updated to use the
  current color map for lookups. These functions are used internally to
  convert integers to a RGB triplet, but they are also available in
  `plot/utils`.

* The `plot-foreground` and `plot-background` parameters have been changed to
  "black" and "white" respectively, this was done to avoid drawing the axes
  and text in the first color of the pen color map and the background in the
  first color of the brush color map.

* `color-map-names` is a new function which returns the list of color map
  names as a list of symbols.

* `color-map-size` is a function which returns the number of colors in a color
  map.  Note that `->pen-color` and `->brush-color` will cycle the colors in
  the color map, so any integer argument is valid for these functions.

* `register-color-map` allows registering new color maps.

The last three functions are available in the `plot/utils` module, but not
`plot`.  This is consistent with the availability of `->pen-color`.

The following color map names are also added, they are based on the matplotlib
and ggplot2 color maps and use the same names: dark2, paired, pastel1,
pastel2, set1, set2, set3, tab10, tab10n, tab20, tab20b and tab20c.

The default plot colors are not changed with this commit, if the user does not
specify a color map using `plot-pen-color-map` or `plot-brush-color-map`, the
plot will use the same colors as before.

Some functions have been moved to different files, this was done to avoid
circular dependencies.