Setup CI for the plot package (#60)
This sets up a CI build for the plot package, scheduled to run for pushes and pull requests. Currently it only runs for Racket version 7.7 and it does not run the tests. Tests need are currently run manually and their output visually inspected.

It is inspired by the web-server CI build, but uses a different strategy for the build:

    minimal racket is installed
    a package catalog is setup for the packages in the plot repository
    this package catalog is added to the default racket catalogs
    the plot package is installed -- this will fetch packages form the local file system and build them, and will fetch all dependencies from the main racket catalog.