Setup Alex as responsible for the tests
Update info.rkt for plot-test to setup Alex as the person responsible for the
plot tests as he is the package maintainer (drdr.racket-lang.org will email
him on failed tests).

Also enable the plot tests in the Github Actions for the package and add
missing "contract-profile" dependency for plot-test.