Implement new method for writing plot tests
This commit introduces the a new way of writing plot tests where a sample plot
is recorded and compared against the current rendering of the plot.  These
types of tests have the ability to automatically detect changes in the way the
plots are rendered.

This commit only adds tests for changes in #5 and #59, since these were
recently merged and had no tests, as well as #55 since this pull request used
3D plots, and I wanted to validate the method for 3D plots as well.

Instead of comparing bitmap or vector images, plots are rendered using
`plot/dc` or `plot3d/dc` to a special `record-dc%` object and the draw steps
are written to a file. When running the tests, the new set of draw steps are
generated and compared to the saved ones.

The `record-dc%` used for drawing is updated to report consistent values for
text dimensions, regardless of the font -- this allows the data files to be
used across different platforms which would generate slightly different plot
images due to different fonts being available.

In addition to this, a sample PNG image is saved for each plot -- this allows
the user to manually run the tests and having something to compare the result.

The verification steps are automated in `check-draw-steps` and
`check-draw-steps3d` such that, when a verification fails, a new set of draw
steps are written to file, as well as a new sample image.

The GitHub Actions `ci.yml` has been updated to upload these files as an
artifact, so we have the data to analyze failed runs on GitHub Actions. The
test will work fine without this feature, so they can be run on DrDr or the
users computer.