Allow pictures to be used as legend entries in plots
Updated 2D and 3D plot renderers to accept a picture (from the pict library)
or a string as the label used in the plot legend.

The original request was to allow LaTeX notation for the labels, this
changeset will enable doing that (and more) by installing the additional
`latex-pict` package and using the pictures produced by that package as
labels.

When updated the code which draws the legend, several considerations were
made:

* the `get-text-extent` method of `plot-device%` accepts a pict as an argument
  and will return the picture dimensions in that case.  The method was not
  renamed, however.

* the font used for text legend entries determines the inset (inside border)
  of the legend, gap between entries and length of the "sample" line.  This
  was essentially unchanged behavior and using the pict dimensions for this
  purpose results in bad looking legends, where the inset is too small, the
  sample line too long and gap too big.

Also updated some of the examples in the documentation to replace x^2 with x²,
to illustrate that Unicode can be used for the string labels.

See also

* https://github.com/racket/plot/issues/58
* https://github.com/racket/racket/issues/3321