Fix handling of empty selection in the 2D plot snip
2D plot snips allow the user to select a region and zoom the plot to that
region -- Sometimes, when the user just clicks on the snip, the system
interprets this as a selection and the snip attempts to zoom to an empty
region.  Empty regions have +nan.0 values, and the code fails with an
exception when it attempts to covert the +nan.0 to an exact number.