Allow pictures to be used as title and axis labels in plots
Updated the 2D and 3D plot procedures to accept a picture (from the pict
library) or a string as the label for the `#:title` and`#:{x,y,z}-label``
keywords, as well as the corresponing parameters, `plot-title` and
`plot-{x,y,z}-label`.

The original request was to allow LaTeX notation for the labels, this
changeset will enable doing that (and more) by installing the additional
`latex-pict` package and using the pictures produced by that package as
labels.

The folowing caveats apply for the implementation:

* the label position for the Y axis **is not** rotated when it is a picture,
  only when it is a string.  This was done to allow full flexibility to users
  who chose to use pictures -- they can easily rotate the picture themselves.
  If the picture would be rotated by the plot package, the user might have to
  supply a rotated picture to have a horizontal picture for the y label, and
  this would result in a double rotation -- depending on the type of graphics
  used for the picture, this might create undesirable artifacts.

* the font used for axis labels and title determines the gap between the label
  and the plot area and axis themselves.

A bug was fixed, where passing incorect parameters to plot procedure could
cause Racket to crash, an explanation for the fix is put as a comment in the
"plot-gui-lib/plot/private/gui/plot2d.rkt" file.

See also:

* https://github.com/racket/plot/issues/58
* https://github.com/racket/racket/issues/3321
* https://github.com/racket/plot/pull/64