Separate legend entry creation from plot rendering (#70, #49)
The various plot renderer procedures for 2D and 3D plots used to render their
data than return a legend entry to be used for constructing the legend.  This
means that the plot library can only know about the list of legend entries
after the data is drawn on the plot.

The `renderer2d` and `renderer3d` structures have been updated to add a slot
for the legend entry and all renders have been updated so the legend entry is
available without rendering the actual data.

This was done so that legend entries can be drawn outside of the plot area,
see #49 for all the commits related to that functionality.