Moved `set-message-center` call to `start-update-thread`.
When a 3d plot snip is resized (e.g when using `plot3d-frame`), a new position
for the "altitude angle" message needs to be computed.  However, since the
message is in the middle of the plot area, the new message location cannot be
set in the `resize` method, as we don't have a new plot area just yet.

Instead, the `set-message-center` call has to be delayed until a new area is
created, which is in `start-update-thread`.