fix CI build to use packages from the repository (#87)
The build workflow would setup the package catalog as current user, but
run the build as root using sudo.  This meant that the build used a
different package catalog and plot packages were retrieved from the
package catalog, instead of being locally build, effectively defeating
the CI process.  This probably worked in the original setup because all
commands were run as "root", but broke when updating to setup-racket
0.10.

Updated the CI workflow to run all racket commands as the current user
and also updated setup-racket to 1.1 and Racket fto 8.0