New box-and-whisker renderer and updated violin renderer
Add a new `box-and-whisker` renderer and update the `violin` renderer to use a
`#:width` parameter instead of the `#:skip` and `#:gap` arguments.  Also
violin renderer now uses a single set of line parameters (color, width and
style) instead of two.

Updated tests and documentation to provide a more comprehensive example.