Remove unused code in glyph drawing code
The original intent of the `when` block was to remove the line around the
glyph when the line and fill color were the same. However, the `eq?` test was
always false in Typed Racket, so the code never executed.

The `eq?` test is now fixed in
https://github.com/racket/typed-racket/pull/1358, but as a result, the visual
layout of plots would change, making glyphs smaller, since the outline is not
drawn anymore.

To keep the backwards compatibility of the plot look-and-feel, the entire
`when` block was removed.